Cybersecurity Incident Investigation and Analysis
This project focuses on a cybersecurity incident involving a suspicious file download at a financial services company. It provides a step-by-step guide for investigating the incident, using tools like SHA256 hashing and VirusTotal for IoC discovery. The goal is to analyze the file's maliciousness and understand associated indicators of compromise (IoCs) using the Pyramid of Pain framework.